+++ innofw/core/datamodules/lightning_datamodules/__init__.py
@@ -1,5 +1,5 @@
 from .image_folder_dm import ImageLightningDataModule
 from .qsar_dm import QsarSelfiesDataModule
-from .segmentation_hdf5_dm import HDF5LightningDataModule
+from .semantic_segmentation.segmentation_hdf5_dm import HDF5LightningDataModule
 
 # from .detection import YOLOv5DataModule
